requirements.txt
streamlit
pandas
feedparser
tqdm
streamlit
pandas
feedparser
requests
arxiv
scikit-learn
sentence-transformers
faiss-cpu